Various changes/cleanup
Take a look at evaluation function, see if we can add things. Changed
some conditionals in there. Added field to player that contains both
current piece symbol and boolean variable that indicates which side we
are.
Added scores to BoardGameMoves, let's think about how we want to pass
the score/move back in the negamax call. Began work on abNegamax.