some cleanup/eval function
Hey I figured out why it didn't work with my 'W' and 'B' chars. In the
BoardGameState.h file, the constant hasn't been defined yet. It will be
defined from the parameters of the game. If you notice, it hasn't even
defined the number of ROWS and COLS yet. So actually I think it might be
ok to use 'W' and 'B'...