Some fixes
I thought this fixed all the issues with the negamax algorithm, but
there are still problems we need to tackle. For example, bt2 is solved
correctly when depth is 1 or 2, but for some reason with depth of 3 and
up it returns a wrong solution. But at least it takes the winning move
at times when available, but sometimes it doesn't! Very confusing and
irksome. We'll have to do serious work with this...