Penalize Deeper wins/more efficient eval function
I penalized the deeper wins, in a way forcing the shortest win. This
resolves the issue of bt2 playing incorrectly.
I also made some efficiency changes to the eval function, we evaluate if
it's a winning position right away. Thus, if it is a winning position,
no need to loop through board and evaluate all pieces.

MERGE BACK INTO MASTER IF YOU LIKE CHANGES